Clean up Store declaration
